Add files via upload
Every bank account is described by its Account Number and Account Name. Model a class BankAccount that can be 
instantiated in a class Main using a parameterized constructor for Account Number and Account Name. 
Additionally, every bank account has its account balance which is by default equivalent to zero and will just be 
updated whenever a bank account holder tries to deposit or withdraw a valid amount. Valid amount is any nonnegative or non-zero number. 
Your task is to define the class needed to execute the program. Carefully implement encapsulation and only expose 
those methods which are needed to successfully perform the simulation of transacting in an ATM Machine using 
the deposit and withdraw operations. 
A menu-driven program is required with a menu for Deposit, Withdraw and Exit. The program should display the 
Account Details of the bank account holder for every selected operation. The program will terminate after the user 
selects the EXIT Menu.